+++ noodles-sam/src/header/program.rs
@@ -172,13 +172,13 @@ impl Program {
     ///
     /// let program = Program::builder()
     ///     .set_id("pg0")
-    ///     .insert(Tag::Other(String::from("zn")), String::from("noodles"))
+    ///     .insert(Tag::Other([b'z', b'n']), String::from("noodles"))
     ///     .build()?;
     ///
     /// let fields = program.fields();
     /// assert_eq!(fields.len(), 1);
     /// assert_eq!(
-    ///     fields.get(&Tag::Other(String::from("zn"))),
+    ///     fields.get(&Tag::Other([b'z', b'n'])),
     ///     Some(&String::from("noodles"))
     /// );
     ///