+++ noodles-sam/src/header/program/tag.rs
@@ -1,6 +1,13 @@
 //! SAM header program tag.
 
-use std::{error, fmt, str::FromStr};
+use std::{
+    convert::TryFrom,
+    error,
+    fmt::{self, Write},
+    str::FromStr,
+};
+
+const LENGTH: usize = 2;
 
 /// A SAM header program tag.
 #[derive(Clone, Debug, Hash, Eq, PartialEq)]
@@ -18,26 +25,29 @@ pub enum Tag {
     /// Program version (`VN`).
     Version,
     /// Any other program tag.
-    Other(String),
+    Other([u8; LENGTH]),
 }
 
-impl AsRef<str> for Tag {
-    fn as_ref(&self) -> &str {
+impl AsRef<[u8; LENGTH]> for Tag {
+    fn as_ref(&self) -> &[u8; LENGTH] {
         match self {
-            Self::Id => "ID",
-            Self::Name => "PN",
-            Self::CommandLine => "CL",
-            Self::PreviousId => "PP",
-            Self::Description => "DS",
-            Self::Version => "VN",
-            Self::Other(s) => s,
+            Self::Id => b"ID",
+            Self::Name => b"PN",
+            Self::CommandLine => b"CL",
+            Self::PreviousId => b"PP",
+            Self::Description => b"DS",
+            Self::Version => b"VN",
+            Self::Other(tag) => tag,
         }
     }
 }
 
 impl fmt::Display for Tag {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.as_ref())
+        let bytes = self.as_ref();
+        f.write_char(char::from(bytes[0]))?;
+        f.write_char(char::from(bytes[1]))?;
+        Ok(())
     }
 }
 
@@ -65,21 +75,28 @@ impl FromStr for Tag {
     type Err = ParseError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "" => Err(ParseError::Empty),
-            "ID" => Ok(Self::Id),
-            "PN" => Ok(Self::Name),
-            "CL" => Ok(Self::CommandLine),
-            "PP" => Ok(Self::PreviousId),
-            "DS" => Ok(Self::Description),
-            "VN" => Ok(Self::Version),
-            _ => {
-                if s.len() == 2 {
-                    Ok(Self::Other(s.into()))
-                } else {
-                    Err(ParseError::Invalid)
-                }
-            }
+        let bytes = s.as_bytes();
+
+        match bytes.len() {
+            0 => Err(ParseError::Empty),
+            2 => Self::try_from([bytes[0], bytes[1]]),
+            _ => Err(ParseError::Invalid),
+        }
+    }
+}
+
+impl TryFrom<[u8; LENGTH]> for Tag {
+    type Error = ParseError;
+
+    fn try_from(b: [u8; LENGTH]) -> Result<Self, Self::Error> {
+        match &b {
+            b"ID" => Ok(Self::Id),
+            b"PN" => Ok(Self::Name),
+            b"CL" => Ok(Self::CommandLine),
+            b"PP" => Ok(Self::PreviousId),
+            b"DS" => Ok(Self::Description),
+            b"VN" => Ok(Self::Version),
+            _ => Ok(Self::Other(b)),
         }
     }
 }
@@ -96,7 +113,7 @@ mod tests {
         assert_eq!(Tag::PreviousId.to_string(), "PP");
         assert_eq!(Tag::Description.to_string(), "DS");
         assert_eq!(Tag::Version.to_string(), "VN");
-        assert_eq!(Tag::Other(String::from("ND")).to_string(), "ND");
+        assert_eq!(Tag::Other([b'N', b'D']).to_string(), "ND");
     }
 
     #[test]
@@ -107,7 +124,7 @@ mod tests {
         assert_eq!("PP".parse(), Ok(Tag::PreviousId));
         assert_eq!("DS".parse(), Ok(Tag::Description));
         assert_eq!("VN".parse(), Ok(Tag::Version));
-        assert_eq!("ND".parse(), Ok(Tag::Other(String::from("ND"))));
+        assert_eq!("ND".parse(), Ok(Tag::Other([b'N', b'D'])));
 
         assert_eq!("".parse::<Tag>(), Err(ParseError::Empty));
         assert_eq!("NDL".parse::<Tag>(), Err(ParseError::Invalid));